Fixed MethodNotFound exception when cross compiling with 1.5
git-svn-id: svn://source.pentaho.org/svnkettleroot/mirror/Kettle/trunk@244 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/be/ibridge/kettle/core/value/ValueBigNumber.java
@@ -112,7 +112,7 @@ public class ValueBigNumber implements ValueInterface, Cloneable
 	
 	public void    setInteger(long number)
 	{
-		this.number = new BigDecimal( number );
+		this.number = new BigDecimal( (double) number );
 	}
